global_bench/pyproject.toml
[project]
name = "benchmarker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "docker>=7.1.0",
    "dotenv>=0.9.9",
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "polars>=1.36.1",
    "psutil>=7.1.3",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "requests>=2.32.5",
    "runpod>=1.8.1",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
